==> 01.Data_Preparation.ipynb <==
# Data Preparation
> Author: [Yalim Demirkesen](github.com/demirkeseny) 

> Date: March 2019

In this notebook, the main purpose is to make the data ready. 

First we have a books dataset from [Kaggle](https://www.kaggle.com/zygmunt/goodbooks-10k) called *GoodBooks-10K*. As the name suggests we have 10,000 book information in this data frame. Since in the next stages of the project, we need a description for each book, [Goodreads](https://github.com/goodreads) module is used. In this module, there are detailed information about each book like:

- Author
- Genre
- Release Date
- Average Ratings
- Rating Distribution
- Number of Pages
- Goodreads Link of the Book
- Description

Since `.description` provides a paragraph about each book that we cannot obtain from *GoodBooks-10K*, we will be able to make NLP analysis and compare the books from their descriptions.

# Necessary libraries:
import pandas as pd
import numpy as np
import xmltodict
import requests
import rauth
from distutils.sysconfig import get_python_lib
import pickle
import os
from goodreads import client

# Enter Goodreads API info the first time running this notebook, then remove
# the file so that the credentials is not uploaded to online platform. They
# will be saved in the pkl file.

if not os.path.exists('secret_goodreads_credentials.pkl'):
    goodreads={}
    goodreads['Consumer Key'] = ''
    goodreads['Consumer Secret'] = ''
    goodreads['Access Token'] = ''
    goodreads['Access Token Secret'] = ''
    with open('secret_goodreads_credentials.pkl','wb') as f:
        pickle.dump(goodreads, f)
else:
    goodreads=pickle.load(open('secret_goodreads_credentials.pkl','rb'))

gc = client.GoodreadsClient(goodreads['Consumer Key'], goodreads['Consumer Secret'])
gc.authenticate(goodreads['Access Token'], goodreads['Access Token Secret'])

# Assigning the book from goodreads.com with the ID number of 1. 
book = gc.book(1)

About the book with the ID number 1, every detailed information can be found from the below queries:

book.title

book.authors[0]

book.average_rating

book.description

book.num_pages

book.link

book.publication_date

book.rating_dist

book.ratings_count

book.text_reviews_count

ids = books.book_id.tolist()

ids[0:10]

book = gc.book(ids[0])
print(book.title)
print(book.description)

After Goodreads package is well-understood, we can create a dataframe that includes the book ID, book title and its description.

books_detailed = pd.DataFrame(index = np.arange(len(books)), 
                              columns = ['id','title','description'])

books_detailed.head()

To fill this data frame, we start with the below for-loop. This is essential to come up with the description of each book.

for i in range(len(ids)):
    try:
        book = gc.book(ids[i])
    except:
        pass
    
    books_detailed['id'][i] = ids[i]
    books_detailed['title'][i] = book.title
    books_detailed['description'][i] = book.description

    if i % 250 == 0:
        print('Filling {}th book out of 10000!'.format(i))
        continue

# Checking the number of rows and columns:
books_detailed.shape

# Changing the ID column to string:
books_detailed['id'] = books_detailed['id'].astype(str)

books_detailed.head()

Now we need to merge this dataframe with our extensive book dataset. But first we need to upload the book.csv file.

books = pd.read_csv('./data/books.csv', encoding='utf-8-sig')

books['book_id'] = books['book_id'].astype(str)

# Checking whether books.csv file has also 10,000 rows:
books.shape

# Dropping the unnecessary columns:
books.drop(columns=['id','best_book_id','isbn13'], inplace = True)

books.head()

# Merging two data files on book ID so that we can have also the book descriptions on the new 'books' variable:
books = pd.merge(books, books_detailed, left_on = 'book_id', right_on='id', how='left')

books.head()

# Dropping unnecessary columns:
books.drop(columns = ['work_id','id','title_y'], inplace = True)

books.head()

# Saving the extended book data frame:
books.to_csv('./data/books_extended.csv')
