==> 03.Collaborative_Filtering_Surprise.ipynb <==
# Collaborative Filtering
> Author: [Yalim Demirkesen](github.com/demirkeseny) 

> Date: March 2019

# Necessary libraries:
import pandas as pd
import surprise
from surprise import SVD
from surprise import SVDpp
from surprise import SlopeOne
from surprise import NMF
from surprise import NormalPredictor
from surprise import KNNBaseline
from surprise import KNNBasic
from surprise import KNNWithMeans
from surprise import KNNWithZScore
from surprise import BaselineOnly
from surprise import CoClustering
from surprise.model_selection import train_test_split
from surprise.accuracy import rmse
from surprise import accuracy
from surprise.model_selection import GridSearchCV
from surprise.model_selection import cross_validate
from surprise import Dataset
from surprise import Reader

### User-Based with Surprise

In user-based collaborative filtering we are utilizing a module called Surprise. That enaables us to compare many models in a much shorter period of time. 

ratings = pd.read_csv('./data/ratings_updated.csv')

We upload our dataframe to the surprise using the `Reader` function.

reader = Reader(rating_scale=(1, 5))
data = Dataset.load_from_df(ratings[['user_id', 'book_id', 'rating']], reader)

comparison = []
# Iterate over all algorithms
for algorithm in [SVD(), SVDpp(), SlopeOne(), NMF(), NormalPredictor(), KNNBaseline(), 
                  KNNBasic(), KNNWithMeans(), KNNWithZScore(), BaselineOnly(), CoClustering()]:
    # Performing cross validation with 3-folds
    results = cross_validate(algorithm, data, measures=['RMSE'], cv=3, verbose=False)
    
    # Get results and create a data frame including the algorithm and its error rate and run time
    tmp = pd.DataFrame.from_dict(results).mean(axis=0)
    tmp = tmp.append(pd.Series([str(algorithm).split(' ')[0].split('.')[-1]], index=['Algorithm']))
    comparison.append(tmp)
    
pd.DataFrame(comparison).set_index('Algorithm').sort_values('test_rmse') 

After `cv=3`, we realized that the best test score is obtained with SVDpp but I prefer SVD since the test score is not that very different than SVDpp but it is much faster. After picking the algorithm, I start by splitting the train and testing data frames. Then run the SVD model with gridsearch which enables me to check the most optimal hyperparameters in model.

trainset, testset = train_test_split(data, test_size=0.20)

param_grid = {'n_factors': [10, 20, 30, 40], 
              'n_epochs': [30, 35, 40], 
              'lr_all': [0.001, 0.003, 0.005, 0.008],
              'reg_all': [0.08, 0.1, 0.15]}

gs = GridSearchCV(SVD, param_grid, measures=['rmse', 'mae'], cv=3)

gs.fit(data)

algo = gs.best_estimator['rmse']
print(gs.best_score['rmse'])
print(gs.best_params['rmse'])

cross_validate(algo, data, measures=['RMSE', 'MAE'], cv=5, verbose=True)

After running our gridsearch, we realized that the best parameters are given above. As next step we build our final model with these parameters.

algo = SVD(n_factors=40, n_epochs=40, lr_all=0.008, reg_all=0.1)
algo.fit(trainset)
test_pred = algo.test(testset)
print("SVD : Test Set")
accuracy.rmse(test_pred, verbose=True)

### Evaluating the Surprise Module

In order to evaluate how our model operates, we generate predicted score for each user in the test set and estimate what they would have scored for certain books. In order to start with that evaluation, I create a data frame displaying the real rating and estimated rating of the user.

test_pred = pd.DataFrame(test_pred)
test_pred.drop(columns = ['details'], inplace = True)
test_pred.columns = ['user_id','book_id','real_rating','est_rating']
test_pred['error'] = abs(test_pred.est_rating - test_pred.real_rating)

test_pred.head()

As a next step, I will check the performance of our estimators.

books_ext = pd.read_csv('./data/books_extended.csv', encoding='utf-8-sig')

books = pd.read_csv('./data/books.csv', encoding='utf-8-sig')

Merging these two datasets so that we can have the most extensive dataset about our books:

book = pd.merge(books[['id','book_id']], books_ext,
                      on='book_id', how = 'inner')

# Dropping the unnecessary columns:
book.drop(columns = ['book_id','Unnamed: 0', 'image_url', 'small_image_url'], 
           inplace = True)

# Just to have the same column names, I change the column 'id' to 'book_id':
book.rename(columns={'id':'book_id'}, inplace = True)

book.head()

Now we will compare the rating given by the user and what out model predicts. In order to do that, we will merge our predictions dataframe and books dataframe. 

main = pd.merge(test_pred, book[['book_id','title_x']], 
                on = 'book_id', how = 'inner')

main.columns.tolist()

main.head()

Just to provide an example, below is the performance of our model for the user `2514`!

main[main['user_id'] == 2514].sort_values(by=['error'])[['user_id','title_x','real_rating','est_rating','error']].reset_index(drop=True)

# Filtering one random user out of all users:
user_interest = main[main['user_id'] == 12874]

# Sorting the dataframe in respect to the error rate and taking the first 45:
user_interest = user_interest.sort_values(by=['error'])[['user_id','title_x','real_rating','est_rating','error']].head(45)

# I reset the index so that user ID also becomes a column
user_interest = user_interest.reset_index(drop=True)

# Then round the rating such that there is only one digit after the comma.
user_interest.est_rating = user_interest.est_rating.round(1)

# Renaming the columns:
user_interest.columns = ['USER ID','BOOK TITLE', 'ACT RATING','EST RATING', 'ERROR']

# Taking the ones with error rate less than 0.5:
user_interest = user_interest[user_interest['ERROR'] < 0.5]

# Sorting by estimated rating so that the books that highly be interesting for the user will be displayed:
user_interest.sort_values(by=['EST RATING'], ascending=False)

Utilizing plotly, we can generate a dynamic graph to see the residuals on the user `12874`.

import plotly.plotly as py
import plotly.graph_objs as go

x = user_interest.sample(n=12, random_state=1)['BOOK TITLE'].tolist()
y = user_interest.sample(n=12, random_state=1)['ACT RATING'].tolist()
y2 = user_interest.sample(n=12, random_state=1)['EST RATING'].tolist()

actual = go.Bar(
    x=x,
    y=y,
    text=y,
    textposition = 'auto',
    marker=dict(
        color='rgb(158,202,225)',
        line=dict(
            color='rgb(8,48,107)',
            width=0.5),
        ),
    opacity=0.6
)

predicted = go.Bar(
    x=x,
    y=y2,
    text=y2,
    textposition = 'auto',
    marker=dict(
        color='rgb(58,200,225)',
        line=dict(
            color='rgb(8,48,107)',
            width=1.5),
        ),
    opacity=0.6
)

data = [actual,predicted]

py.iplot(data, filename='grouped-bar-direct-labels')

# Building its function:
def top_choice(user_identification, top):
    top_ch = main[main['user_id'] == user_identification].sort_values(by=['error'])[['user_id',
                                                              'title_x',
                                                              'real_rating',
                                                              'est_rating',
                                                              'error']].reset_index(drop=True)
    top_ch = top_ch.sort_values(by=['est_rating'], ascending = False).head(top)
    return(top_ch)

# Thanks to the above function, we can generate the books that the user 
# would enjoy most according to our recommendation engine:
top_choice(12874, 5)
